fix: Ensure `FrameValue` rejects `float`

+++ packages/vgplot/spec-python/src/mosaic_spec/_gen/transform.py
@@ -166,7 +166,7 @@ class BinOptions(_BinOptions, closed=True):
     """Bin transform options."""
 
 
-FrameValue = TypeAliasType("FrameValue", IntervalTransform | float | None)
+FrameValue = TypeAliasType("FrameValue", IntervalTransform | int | None)
 ColumnTransform = TypeAliasType(
     "ColumnTransform",
     Bin | Centroid | CentroidX | CentroidY | Column | DateDay | DateMonth | DateMonthDay | GeoJSON,
